🔧 turned off persistent sessions
this is typically turned off by default but Spring devtools turns it on.

this was causing an issue where users would stay logged in when the program is stopped and restarted, but since the database gets wiped the logged in user technically did not exist which just caused all sorts of issues.